daemon: move general api to api_general*.go
This moves general APIs (that are not specific to some entity
or concept in snapd) to api_general.go and api_general_test.go
from api.go/api_test.go.
