+++ daemon/api_test.go
@@ -45,7 +45,6 @@ import (
 	"github.com/snapcore/snapd/arch"
 	"github.com/snapcore/snapd/asserts"
 	"github.com/snapcore/snapd/asserts/assertstest"
-	"github.com/snapcore/snapd/boot"
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/interfaces"
@@ -64,7 +63,6 @@ import (
 	"github.com/snapcore/snapd/overlord/servicestate"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
-	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/sandbox"
 	"github.com/snapcore/snapd/snap"
 	"github.com/snapcore/snapd/snap/channel"
@@ -507,301 +505,6 @@ func (s *apiSuite) TestListIncludesAll(c *check.C) {
 
 	c.Check(found, check.Equals, len(api),
 		check.Commentf(`At a glance it looks like you've not added all the Commands defined in api to the api list.`))
-}
-
-func (s *apiSuite) TestRootCmd(c *check.C) {
-	// check it only does GET
-	c.Check(rootCmd.PUT, check.IsNil)
-	c.Check(rootCmd.POST, check.IsNil)
-	c.Assert(rootCmd.GET, check.NotNil)
-
-	rec := httptest.NewRecorder()
-	c.Check(rootCmd.Path, check.Equals, "/")
-
-	rootCmd.GET(rootCmd, nil, nil).ServeHTTP(rec, nil)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.HeaderMap.Get("Content-Type"), check.Equals, "application/json")
-
-	expected := []interface{}{"TBD"}
-	var rsp resp
-	c.Assert(json.Unmarshal(rec.Body.Bytes(), &rsp), check.IsNil)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Result, check.DeepEquals, expected)
-}
-
-func mockSystemdVirt(newVirt string) (restore func()) {
-	oldVirt := systemdVirt
-	systemdVirt = newVirt
-	return func() { systemdVirt = oldVirt }
-}
-
-func (s *apiSuite) TestSysInfo(c *check.C) {
-	// check it only does GET
-	c.Check(sysInfoCmd.PUT, check.IsNil)
-	c.Check(sysInfoCmd.POST, check.IsNil)
-	c.Assert(sysInfoCmd.GET, check.NotNil)
-
-	rec := httptest.NewRecorder()
-	c.Check(sysInfoCmd.Path, check.Equals, "/v2/system-info")
-
-	d := s.daemon(c)
-	d.Version = "42b1"
-
-	// set both legacy and new refresh schedules. new one takes priority
-	st := d.overlord.State()
-	st.Lock()
-	tr := config.NewTransaction(st)
-	tr.Set("core", "refresh.schedule", "00:00-9:00/12:00-13:00")
-	tr.Set("core", "refresh.timer", "8:00~9:00/2")
-	tr.Commit()
-	st.Unlock()
-
-	restore := release.MockReleaseInfo(&release.OS{ID: "distro-id", VersionID: "1.2"})
-	defer restore()
-	restore = release.MockOnClassic(true)
-	defer restore()
-	restore = sandbox.MockForceDevMode(true)
-	defer restore()
-	// reload dirs for release info to have effect
-	dirs.SetRootDir(dirs.GlobalRootDir)
-	restore = mockSystemdVirt("magic")
-	defer restore()
-
-	buildID := "this-is-my-build-id"
-	restore = MockBuildID(buildID)
-	defer restore()
-
-	sysInfoCmd.GET(sysInfoCmd, nil, nil).ServeHTTP(rec, nil)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.HeaderMap.Get("Content-Type"), check.Equals, "application/json")
-
-	expected := map[string]interface{}{
-		"series":  "16",
-		"version": "42b1",
-		"os-release": map[string]interface{}{
-			"id":         "distro-id",
-			"version-id": "1.2",
-		},
-		"build-id":   buildID,
-		"on-classic": true,
-		"managed":    false,
-		"locations": map[string]interface{}{
-			"snap-mount-dir": dirs.SnapMountDir,
-			"snap-bin-dir":   dirs.SnapBinariesDir,
-		},
-		"refresh": map[string]interface{}{
-			// only the "timer" field
-			"timer": "8:00~9:00/2",
-		},
-		"confinement":      "partial",
-		"sandbox-features": map[string]interface{}{"confinement-options": []interface{}{"classic", "devmode"}},
-		"architecture":     arch.DpkgArchitecture(),
-		"virtualization":   "magic",
-		"system-mode":      "run",
-	}
-	var rsp resp
-	c.Assert(json.Unmarshal(rec.Body.Bytes(), &rsp), check.IsNil)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	// Ensure that we had a kernel-verrsion but don't check the actual value.
-	const kernelVersionKey = "kernel-version"
-	c.Check(rsp.Result.(map[string]interface{})[kernelVersionKey], check.Not(check.Equals), "")
-	delete(rsp.Result.(map[string]interface{}), kernelVersionKey)
-	c.Check(rsp.Result, check.DeepEquals, expected)
-}
-
-func (s *apiSuite) TestSysInfoLegacyRefresh(c *check.C) {
-	rec := httptest.NewRecorder()
-
-	d := s.daemon(c)
-	d.Version = "42b1"
-
-	restore := release.MockReleaseInfo(&release.OS{ID: "distro-id", VersionID: "1.2"})
-	defer restore()
-	restore = release.MockOnClassic(true)
-	defer restore()
-	restore = sandbox.MockForceDevMode(true)
-	defer restore()
-	restore = mockSystemdVirt("kvm")
-	defer restore()
-	// reload dirs for release info to have effect
-	dirs.SetRootDir(dirs.GlobalRootDir)
-
-	// set the legacy refresh schedule
-	st := d.overlord.State()
-	st.Lock()
-	tr := config.NewTransaction(st)
-	tr.Set("core", "refresh.schedule", "00:00-9:00/12:00-13:00")
-	tr.Set("core", "refresh.timer", "")
-	tr.Commit()
-	st.Unlock()
-
-	// add a test security backend
-	err := d.overlord.InterfaceManager().Repository().AddBackend(&ifacetest.TestSecurityBackend{
-		BackendName:             "apparmor",
-		SandboxFeaturesCallback: func() []string { return []string{"feature-1", "feature-2"} },
-	})
-	c.Assert(err, check.IsNil)
-
-	buildID := "this-is-my-build-id"
-	restore = MockBuildID(buildID)
-	defer restore()
-
-	sysInfoCmd.GET(sysInfoCmd, nil, nil).ServeHTTP(rec, nil)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.HeaderMap.Get("Content-Type"), check.Equals, "application/json")
-
-	expected := map[string]interface{}{
-		"series":  "16",
-		"version": "42b1",
-		"os-release": map[string]interface{}{
-			"id":         "distro-id",
-			"version-id": "1.2",
-		},
-		"build-id":   buildID,
-		"on-classic": true,
-		"managed":    false,
-		"locations": map[string]interface{}{
-			"snap-mount-dir": dirs.SnapMountDir,
-			"snap-bin-dir":   dirs.SnapBinariesDir,
-		},
-		"refresh": map[string]interface{}{
-			// only the "schedule" field
-			"schedule": "00:00-9:00/12:00-13:00",
-		},
-		"confinement": "partial",
-		"sandbox-features": map[string]interface{}{
-			"apparmor":            []interface{}{"feature-1", "feature-2"},
-			"confinement-options": []interface{}{"classic", "devmode"}, // we know it's this because of the release.Mock... calls above
-		},
-		"architecture":   arch.DpkgArchitecture(),
-		"virtualization": "kvm",
-		"system-mode":    "run",
-	}
-	var rsp resp
-	c.Assert(json.Unmarshal(rec.Body.Bytes(), &rsp), check.IsNil)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	const kernelVersionKey = "kernel-version"
-	delete(rsp.Result.(map[string]interface{}), kernelVersionKey)
-	c.Check(rsp.Result, check.DeepEquals, expected)
-}
-
-func (s *apiSuite) testSysInfoSystemMode(c *check.C, mode string) {
-	c.Assert(mode != "", check.Equals, true, check.Commentf("mode is unset for the test"))
-	rec := httptest.NewRecorder()
-
-	restore := release.MockReleaseInfo(&release.OS{ID: "distro-id", VersionID: "1.2"})
-	defer restore()
-	restore = release.MockOnClassic(false)
-	defer restore()
-	restore = sandbox.MockForceDevMode(false)
-	defer restore()
-	restore = mockSystemdVirt("")
-	defer restore()
-
-	// reload dirs for release info to have effect on paths
-	dirs.SetRootDir(dirs.GlobalRootDir)
-
-	// mock the modeenv file
-	m := boot.Modeenv{
-		Mode:           mode,
-		RecoverySystem: "20191127",
-	}
-	err := m.WriteTo("")
-	c.Assert(err, check.IsNil)
-
-	d := s.daemon(c)
-	d.Version = "42b1"
-
-	// add a test security backend
-	err = d.overlord.InterfaceManager().Repository().AddBackend(&ifacetest.TestSecurityBackend{
-		BackendName:             "apparmor",
-		SandboxFeaturesCallback: func() []string { return []string{"feature-1", "feature-2"} },
-	})
-	c.Assert(err, check.IsNil)
-
-	buildID := "this-is-my-build-id"
-	restore = MockBuildID(buildID)
-	defer restore()
-
-	sysInfoCmd.GET(sysInfoCmd, nil, nil).ServeHTTP(rec, nil)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.HeaderMap.Get("Content-Type"), check.Equals, "application/json")
-
-	expected := map[string]interface{}{
-		"series":  "16",
-		"version": "42b1",
-		"os-release": map[string]interface{}{
-			"id":         "distro-id",
-			"version-id": "1.2",
-		},
-		"build-id":   buildID,
-		"on-classic": false,
-		"managed":    false,
-		"locations": map[string]interface{}{
-			"snap-mount-dir": dirs.SnapMountDir,
-			"snap-bin-dir":   dirs.SnapBinariesDir,
-		},
-		"refresh": map[string]interface{}{
-			"timer": "00:00~24:00/4",
-		},
-		"confinement": "strict",
-		"sandbox-features": map[string]interface{}{
-			"apparmor":            []interface{}{"feature-1", "feature-2"},
-			"confinement-options": []interface{}{"devmode", "strict"}, // we know it's this because of the release.Mock... calls above
-		},
-		"architecture": arch.DpkgArchitecture(),
-		"system-mode":  mode,
-	}
-	var rsp resp
-	c.Assert(json.Unmarshal(rec.Body.Bytes(), &rsp), check.IsNil)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	const kernelVersionKey = "kernel-version"
-	delete(rsp.Result.(map[string]interface{}), kernelVersionKey)
-	c.Check(rsp.Result, check.DeepEquals, expected)
-}
-
-func (s *apiSuite) TestSysInfoSystemModeRun(c *check.C) {
-	s.testSysInfoSystemMode(c, "run")
-}
-
-func (s *apiSuite) TestSysInfoSystemModeRecover(c *check.C) {
-	s.testSysInfoSystemMode(c, "recover")
-}
-
-func (s *apiSuite) TestSysInfoSystemModeInstall(c *check.C) {
-	s.testSysInfoSystemMode(c, "install")
-}
-func (s *apiSuite) TestSysInfoIsManaged(c *check.C) {
-	s.daemon(c)
-
-	st := s.d.overlord.State()
-	st.Lock()
-	_, err := auth.NewUser(st, "someuser", "mymail@test.com", "macaroon", []string{"discharge"})
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	req, err := http.NewRequest("GET", "/v2/system-info", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := sysInfo(sysInfoCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Result.(map[string]interface{})["managed"], check.Equals, true)
-}
-
-func (s *apiSuite) TestSysInfoWorksDegraded(c *check.C) {
-	s.daemon(c)
-
-	s.d.SetDegradedMode(fmt.Errorf("some error"))
-
-	req, err := http.NewRequest("GET", "/v2/system-info", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := sysInfo(sysInfoCmd, req, nil).(*resp)
-	c.Check(rsp.Status, check.Equals, 200)
 }
 
 func (s *apiSuite) TestLoginUser(c *check.C) {
@@ -5164,362 +4867,6 @@ func (s *apiSuite) TestUnsupportedInterfaceAction(c *check.C) {
 	})
 }
 
-func setupChanges(st *state.State) []string {
-	chg1 := st.NewChange("install", "install...")
-	chg1.Set("snap-names", []string{"funky-snap-name"})
-	t1 := st.NewTask("download", "1...")
-	t2 := st.NewTask("activate", "2...")
-	chg1.AddAll(state.NewTaskSet(t1, t2))
-	t1.Logf("l11")
-	t1.Logf("l12")
-	chg2 := st.NewChange("remove", "remove..")
-	t3 := st.NewTask("unlink", "1...")
-	chg2.AddTask(t3)
-	t3.SetStatus(state.ErrorStatus)
-	t3.Errorf("rm failed")
-
-	return []string{chg1.ID(), chg2.ID(), t1.ID(), t2.ID(), t3.ID()}
-}
-
-func (s *apiSuite) TestStateChangesDefaultToInProgress(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	setupChanges(st)
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.HasLen, 1)
-
-	res, err := rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-
-	c.Check(string(res), check.Matches, `.*{"id":"\w+","kind":"install","summary":"install...","status":"Do","tasks":\[{"id":"\w+","kind":"download","summary":"1...","status":"Do","log":\["2016-04-21T01:02:03Z INFO l11","2016-04-21T01:02:03Z INFO l12"],"progress":{"label":"","done":0,"total":1},"spawn-time":"2016-04-21T01:02:03Z"}.*`)
-}
-
-func (s *apiSuite) TestStateChangesInProgress(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	setupChanges(st)
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes?select=in-progress", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.HasLen, 1)
-
-	res, err := rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-
-	c.Check(string(res), check.Matches, `.*{"id":"\w+","kind":"install","summary":"install...","status":"Do","tasks":\[{"id":"\w+","kind":"download","summary":"1...","status":"Do","log":\["2016-04-21T01:02:03Z INFO l11","2016-04-21T01:02:03Z INFO l12"],"progress":{"label":"","done":0,"total":1},"spawn-time":"2016-04-21T01:02:03Z"}.*],"ready":false,"spawn-time":"2016-04-21T01:02:03Z"}.*`)
-}
-
-func (s *apiSuite) TestStateChangesAll(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	setupChanges(st)
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes?select=all", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.HasLen, 2)
-
-	res, err := rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-
-	c.Check(string(res), check.Matches, `.*{"id":"\w+","kind":"install","summary":"install...","status":"Do","tasks":\[{"id":"\w+","kind":"download","summary":"1...","status":"Do","log":\["2016-04-21T01:02:03Z INFO l11","2016-04-21T01:02:03Z INFO l12"],"progress":{"label":"","done":0,"total":1},"spawn-time":"2016-04-21T01:02:03Z"}.*],"ready":false,"spawn-time":"2016-04-21T01:02:03Z"}.*`)
-	c.Check(string(res), check.Matches, `.*{"id":"\w+","kind":"remove","summary":"remove..","status":"Error","tasks":\[{"id":"\w+","kind":"unlink","summary":"1...","status":"Error","log":\["2016-04-21T01:02:03Z ERROR rm failed"],"progress":{"label":"","done":1,"total":1},"spawn-time":"2016-04-21T01:02:03Z","ready-time":"2016-04-21T01:02:03Z"}.*],"ready":true,"err":"[^"]+".*`)
-}
-
-func (s *apiSuite) TestStateChangesReady(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	setupChanges(st)
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes?select=ready", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.HasLen, 1)
-
-	res, err := rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-
-	c.Check(string(res), check.Matches, `.*{"id":"\w+","kind":"remove","summary":"remove..","status":"Error","tasks":\[{"id":"\w+","kind":"unlink","summary":"1...","status":"Error","log":\["2016-04-21T01:02:03Z ERROR rm failed"],"progress":{"label":"","done":1,"total":1},"spawn-time":"2016-04-21T01:02:03Z","ready-time":"2016-04-21T01:02:03Z"}.*],"ready":true,"err":"[^"]+".*`)
-}
-
-func (s *apiSuite) TestStateChangesForSnapName(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	setupChanges(st)
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes?for=funky-snap-name&select=all", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.FitsTypeOf, []*changeInfo(nil))
-
-	res := rsp.Result.([]*changeInfo)
-	c.Assert(res, check.HasLen, 1)
-	c.Check(res[0].Kind, check.Equals, `install`)
-
-	_, err = rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-}
-
-func (s *apiSuite) TestStateChangesForSnapNameWithApp(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	chg1 := st.NewChange("service-control", "install...")
-	// as triggered by snap restart lxd.daemon
-	chg1.Set("snap-names", []string{"lxd.daemon"})
-	t1 := st.NewTask("exec-command", "1...")
-	chg1.AddAll(state.NewTaskSet(t1))
-	t1.Logf("foobar")
-
-	st.Unlock()
-
-	// Execute
-	req, err := http.NewRequest("GET", "/v2/changes?for=lxd&select=all", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChanges(stateChangesCmd, req, nil).(*resp)
-
-	// Verify
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.FitsTypeOf, []*changeInfo(nil))
-
-	res := rsp.Result.([]*changeInfo)
-	c.Assert(res, check.HasLen, 1)
-	c.Check(res[0].Kind, check.Equals, `service-control`)
-
-	_, err = rsp.MarshalJSON()
-	c.Assert(err, check.IsNil)
-}
-
-func (s *apiSuite) TestStateChange(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	ids := setupChanges(st)
-	chg := st.Change(ids[0])
-	chg.Set("api-data", map[string]int{"n": 42})
-	st.Unlock()
-	s.vars = map[string]string{"id": ids[0]}
-
-	// Execute
-	req, err := http.NewRequest("POST", "/v2/change/"+ids[0], nil)
-	c.Assert(err, check.IsNil)
-	rsp := getChange(stateChangeCmd, req, nil).(*resp)
-	rec := httptest.NewRecorder()
-	rsp.ServeHTTP(rec, req)
-
-	// Verify
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Result, check.NotNil)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	c.Check(body["result"], check.DeepEquals, map[string]interface{}{
-		"id":         ids[0],
-		"kind":       "install",
-		"summary":    "install...",
-		"status":     "Do",
-		"ready":      false,
-		"spawn-time": "2016-04-21T01:02:03Z",
-		"tasks": []interface{}{
-			map[string]interface{}{
-				"id":         ids[2],
-				"kind":       "download",
-				"summary":    "1...",
-				"status":     "Do",
-				"log":        []interface{}{"2016-04-21T01:02:03Z INFO l11", "2016-04-21T01:02:03Z INFO l12"},
-				"progress":   map[string]interface{}{"label": "", "done": 0., "total": 1.},
-				"spawn-time": "2016-04-21T01:02:03Z",
-			},
-			map[string]interface{}{
-				"id":         ids[3],
-				"kind":       "activate",
-				"summary":    "2...",
-				"status":     "Do",
-				"progress":   map[string]interface{}{"label": "", "done": 0., "total": 1.},
-				"spawn-time": "2016-04-21T01:02:03Z",
-			},
-		},
-		"data": map[string]interface{}{
-			"n": float64(42),
-		},
-	})
-}
-
-func (s *apiSuite) TestStateChangeAbort(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	soon := 0
-	ensureStateSoon = func(st *state.State) {
-		soon++
-	}
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	ids := setupChanges(st)
-	st.Unlock()
-	s.vars = map[string]string{"id": ids[0]}
-
-	buf := bytes.NewBufferString(`{"action": "abort"}`)
-
-	// Execute
-	req, err := http.NewRequest("POST", "/v2/changes/"+ids[0], buf)
-	c.Assert(err, check.IsNil)
-	rsp := abortChange(stateChangeCmd, req, nil).(*resp)
-	rec := httptest.NewRecorder()
-	rsp.ServeHTTP(rec, req)
-
-	// Ensure scheduled
-	c.Check(soon, check.Equals, 1)
-
-	// Verify
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Result, check.NotNil)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	c.Check(body["result"], check.DeepEquals, map[string]interface{}{
-		"id":         ids[0],
-		"kind":       "install",
-		"summary":    "install...",
-		"status":     "Hold",
-		"ready":      true,
-		"spawn-time": "2016-04-21T01:02:03Z",
-		"ready-time": "2016-04-21T01:02:03Z",
-		"tasks": []interface{}{
-			map[string]interface{}{
-				"id":         ids[2],
-				"kind":       "download",
-				"summary":    "1...",
-				"status":     "Hold",
-				"log":        []interface{}{"2016-04-21T01:02:03Z INFO l11", "2016-04-21T01:02:03Z INFO l12"},
-				"progress":   map[string]interface{}{"label": "", "done": 1., "total": 1.},
-				"spawn-time": "2016-04-21T01:02:03Z",
-				"ready-time": "2016-04-21T01:02:03Z",
-			},
-			map[string]interface{}{
-				"id":         ids[3],
-				"kind":       "activate",
-				"summary":    "2...",
-				"status":     "Hold",
-				"progress":   map[string]interface{}{"label": "", "done": 1., "total": 1.},
-				"spawn-time": "2016-04-21T01:02:03Z",
-				"ready-time": "2016-04-21T01:02:03Z",
-			},
-		},
-	})
-}
-
-func (s *apiSuite) TestStateChangeAbortIsReady(c *check.C) {
-	restore := state.MockTime(time.Date(2016, 04, 21, 1, 2, 3, 0, time.UTC))
-	defer restore()
-
-	// Setup
-	d := newTestDaemon(c)
-	st := d.overlord.State()
-	st.Lock()
-	ids := setupChanges(st)
-	st.Change(ids[0]).SetStatus(state.DoneStatus)
-	st.Unlock()
-	s.vars = map[string]string{"id": ids[0]}
-
-	buf := bytes.NewBufferString(`{"action": "abort"}`)
-
-	// Execute
-	req, err := http.NewRequest("POST", "/v2/changes/"+ids[0], buf)
-	c.Assert(err, check.IsNil)
-	rsp := abortChange(stateChangeCmd, req, nil).(*resp)
-	rec := httptest.NewRecorder()
-	rsp.ServeHTTP(rec, req)
-
-	// Verify
-	c.Check(rec.Code, check.Equals, 400)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result, check.NotNil)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	c.Check(body["result"], check.DeepEquals, map[string]interface{}{
-		"message": fmt.Sprintf("cannot abort change %s with nothing pending", ids[0]),
-	})
-}
-
 const validBuyInput = `{
 		  "snap-id": "the-snap-id-1234abcd",
 		  "snap-name": "the snap name",
@@ -7110,61 +6457,6 @@ func (s *apiSuite) TestErrToResponseForRevisionNotAvailable(c *check.C) {
 			Value:   "foo",
 		},
 	})
-}
-
-func (s *apiSuite) testWarnings(c *check.C, all bool, body io.Reader) (calls string, result interface{}) {
-	s.daemon(c)
-
-	oldOK := stateOkayWarnings
-	oldAll := stateAllWarnings
-	oldPending := statePendingWarnings
-	stateOkayWarnings = func(*state.State, time.Time) int { calls += "ok"; return 0 }
-	stateAllWarnings = func(*state.State) []*state.Warning { calls += "all"; return nil }
-	statePendingWarnings = func(*state.State) ([]*state.Warning, time.Time) { calls += "show"; return nil, time.Time{} }
-	defer func() {
-		stateOkayWarnings = oldOK
-		stateAllWarnings = oldAll
-		statePendingWarnings = oldPending
-	}()
-
-	method := "GET"
-	f := warningsCmd.GET
-	if body != nil {
-		method = "POST"
-		f = warningsCmd.POST
-	}
-	q := url.Values{}
-	if all {
-		q.Set("select", "all")
-	}
-	req, err := http.NewRequest(method, "/v2/warnings?"+q.Encode(), body)
-	c.Assert(err, check.IsNil)
-
-	rsp, ok := f(warningsCmd, req, nil).(*resp)
-	c.Assert(ok, check.Equals, true)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Assert(rsp.Result, check.NotNil)
-	return calls, rsp.Result
-}
-
-func (s *apiSuite) TestAllWarnings(c *check.C) {
-	calls, result := s.testWarnings(c, true, nil)
-	c.Check(calls, check.Equals, "all")
-	c.Check(result, check.DeepEquals, []state.Warning{})
-}
-
-func (s *apiSuite) TestSomeWarnings(c *check.C) {
-	calls, result := s.testWarnings(c, false, nil)
-	c.Check(calls, check.Equals, "show")
-	c.Check(result, check.DeepEquals, []state.Warning{})
-}
-
-func (s *apiSuite) TestAckWarnings(c *check.C) {
-	calls, result := s.testWarnings(c, false, bytes.NewReader([]byte(`{"action": "okay", "timestamp": "2006-01-02T15:04:05Z"}`)))
-	c.Check(calls, check.Equals, "ok")
-	c.Check(result, check.DeepEquals, 0)
 }
 
 func (s *apiSuite) TestErrToResponseForChangeConflict(c *check.C) {
